Create My Balance.
This is a small project I made to check my monthly or weekly money balance on a daily balance. This project allows me to keep track of my balance, add to it and withdraw from it. It can also be used by multiple users to keep a track of their respective balances. 
I made this project using Object Oriented Programming, functions and if/else statements in Python.